simplification of overall impl and removal of unused legacy code NXCM-5302 - Protect against staging.xml corruption
