+++ nexus-core/src/main/java/org/sonatype/nexus/util/configuration/ConfigurationHelper.java
@@ -1,0 +1,25 @@
+/*
+ * Sonatype Nexus (TM) Open Source Version
+ * Copyright (c) 2007-2013 Sonatype, Inc.
+ * All rights reserved. Includes the third-party code listed at http://links.sonatype.com/products/nexus/oss/attributions.
+ *
+ * This program and the accompanying materials are made available under the terms of the Eclipse Public License Version 1.0,
+ * which accompanies this distribution and is available at http://www.eclipse.org/legal/epl-v10.html.
+ *
+ * Sonatype Nexus (TM) Professional Version is available from Sonatype, Inc. "Sonatype" and "Sonatype Nexus" are trademarks
+ * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
+ * Eclipse Foundation. All other trademarks are the property of their respective owners.
+ */
+package org.sonatype.nexus.util.configuration;
+
+import java.io.File;
+import java.io.IOException;
+
+/**
+ * Common operations to save configuration files.
+ */
+public interface ConfigurationHelper
+{
+    public <E> void save( E configuration, File configurationFile,
+                          ConfigurationWriter<E> configurationWriter) throws IOException;
+}
