+++ nexus-core/src/main/java/org/sonatype/nexus/util/configuration/ConfigurationReader.java
@@ -10,15 +10,17 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.util.configurationreader;
+package org.sonatype.nexus.util.configuration;
 
 import java.io.IOException;
-import java.io.Writer;
+import java.io.Reader;
 
-public interface ConfigurationWriter<E>
+import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
+
+public interface ConfigurationReader<E>
 {
 
-    void write( Writer fr, E configuration )
-        throws IOException;
+    E read( Reader fr )
+        throws IOException, XmlPullParserException;
 
 }
